+++ NEWS.md
@@ -20,6 +20,8 @@ plumber 1.0.0
 
 ### Breaking changes
 
+* When `plumb()`ing a file (or `plumber$new(file)`), the working directory is set to the file's directory before parsing the file. When running the Plumber API, the working directory will be set to file's directory before running.(#631)
+
 * Plumber's OpenAPI Specification is now defined using
   [OpenAPI 3](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.3.md),
   upgrading from Swagger Specification. (#365)
@@ -31,8 +33,6 @@ plumber 1.0.0
   #328).
 
 * R repository modified to `focal-cran40` using Ubuntu 20.04 LTS for digital ocean provisioning (@meztez, #529)
-
-* When `plumb()`ing a file, the working directory is set to the file's directory. (#631)
 
 
 ### New features